requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
blis==0.7.11
catalogue==2.0.10
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.19.0
confection==0.1.5
cryptography==43.0.1
cymem==2.0.8
Django==5.1.1
djangorestframework==3.15.2
docx2txt==0.8
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.1
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
phonenumbers==8.13.45
pillow==10.4.0
preshed==3.0.9
prompt_toolkit==3.0.47
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.0.8
regex==2024.9.11
requests==2.32.3
resume-parser==0.8.4
rich==13.8.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stemming==1.0.1
thinc==8.2.5
tika==2.6.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0

pyresparser~=1.0.6